requirements.txt
Flask==2.2.2
torch==2.0.1
torchvision==0.15.2
sentence-transformers==2.2.2
Pillow==9.5.0
scikit-learn==1.3.0
git+https://github.com/openai/CLIP.git
